fix(mcp): advertise term/payment enums and split the search test file
Addresses two CodeRabbit findings on the search fan-out commit (d76aa7c00).

The tool schema constrained only provider and lookback_period, so an MCP
client could not discover the valid payment_option/term_years values from the
schema and had to learn them by sending a value the tool rejects. Add the enum
overrides, mirroring ValidatePaymentOption / ValidateTermYears. Those
validators remain the enforcing guard: a client may send anything regardless
of what the schema declares, so this is discoverability, not a new gate.

TestSearchRecommendationsSchemaAdvertisesTermAndPaymentEnums pins both value
sets through the real MCP protocol (ListTools), following the existing Azure
partial-upfront precedent in this file, so the schema cannot drift from the
validators unnoticed. Confirmed to fail when either override is removed.

search_recommendations_test.go had grown to 651 lines, past the 500-line
guideline in CLAUDE.md. Split the AWS reservation fan-out tests into
search_recommendations_fanout_test.go along the existing topical seam,
leaving 473 and 190 lines. The shared fakes stay in the original file and
remain visible to both: same package, different file.
